+++ keyboards/mechwild/mokulua/mirrored/keymaps/default/readme.md
@@ -1,0 +1,1 @@
+# The default keymap for Mokulua using a mirrored right-half
