+++ keyboards/mechwild/mokulua/rules.mk
@@ -1,22 +1,1 @@
-# MCU name
-MCU = atmega32u4
-
-# Bootloader selection
-BOOTLOADER = caterina
-
-# Build Options
-#   change yes to no to disable
-#
-BOOTMAGIC_ENABLE = yes      # Enable Bootmagic Lite
-MOUSEKEY_ENABLE = yes       # Mouse keys
-EXTRAKEY_ENABLE = yes       # Audio control and System control
-CONSOLE_ENABLE = no         # Console for debug
-COMMAND_ENABLE = no         # Commands for debug and configuration
-NKRO_ENABLE = no            # Enable N-Key Rollover
-BACKLIGHT_ENABLE = no       # Enable keyboard backlight functionality
-RGBLIGHT_ENABLE = yes       # Enable keyboard RGB underglow
-AUDIO_ENABLE = no           # Audio output
-ENCODER_ENABLE = yes        # Enable encoder
-OLED_ENABLE = yes           # Enable OLED Screen
-OLED_DRIVER = SSD1306       # Define OLED Driver
-SPLIT_KEYBOARD = yes        # Define split functionality
+DEFAULT_FOLDER = mechwild/mokulua/standard
