+++ keyboards/mechwild/mokulua/standard/standard.h
@@ -7,7 +7,7 @@
 
 #define ___ KC_NO
 
-#define LAYOUT( \
+#define LAYOUT_standard( \
     L00, L01, L02, L03, L04, L05, L06,             R00, R01, R02, R03, R04, R05, R06, \
     L10, L11, L12, L13, L14, L15, L16,               R11, R12, R13, R14,   R15,  R16, \
     L20, L21, L22, L23, L24, L25, L26,             R20, R21, R22, R23, R24, R25, R26, \
@@ -27,3 +27,5 @@
     { R41, R42, R43, R44, R45, R46 }, \
     { R00, R20, R31, R30, R40, ___ }  \
 }
+
+#define LAYOUT       LAYOUT_standard
